+++ @xen-orchestra/web/src/modules/treeview/components/PoolTreeItem.vue
@@ -10,6 +10,19 @@
           variant="secondary"
           size="small"
         />
+        <MenuList placement="bottom-start">
+          <template #trigger="{ open, isOpen }">
+            <UiButtonIcon
+              accent="brand"
+              icon="action:more-actions"
+              variant="tertiary"
+              size="small"
+              :selected="isOpen"
+              @click="open($event)"
+            />
+          </template>
+          <PoolTreeActions :pool="branch.data" />
+        </MenuList>
       </template>
     </UiTreeItemLabel>
     <template v-if="branch.hasChildren" #sublist>
@@ -22,19 +35,22 @@
 </template>
 
 <script lang="ts" setup>
+import PoolTreeActions from '@/modules/pool/components/actions/PoolTreeActions.vue'
 import HostTreeList from '@/modules/treeview/components/HostTreeList.vue'
 import VmTreeList from '@/modules/treeview/components/VmTreeList.vue'
 import type { HostBranch, PoolBranch, VmLeaf } from '@/modules/treeview/types/tree.type.ts'
 import { useXoVmCollection } from '@/modules/vm/remote-resources/use-xo-vm-collection.ts'
+import MenuList from '@core/components/menu/MenuList.vue'
 import VtsTreeItem from '@core/components/tree/VtsTreeItem.vue'
 import VtsTreeList from '@core/components/tree/VtsTreeList.vue'
+import UiButtonIcon from '@core/components/ui/button-icon/UiButtonIcon.vue'
 import UiCounter from '@core/components/ui/counter/UiCounter.vue'
 import UiTreeItemLabel from '@core/components/ui/tree-item-label/UiTreeItemLabel.vue'
 import { vTooltip } from '@core/directives/tooltip.directive.ts'
 import { computed } from 'vue'
 import { useI18n } from 'vue-i18n'
 
-const props = defineProps<{
+const { branch } = defineProps<{
   branch: PoolBranch
 }>()
 
@@ -42,13 +58,11 @@ const { t } = useI18n()
 
 const { runningVms } = useXoVmCollection()
 
-const treeBranches = computed(
-  () => props.branch.children.filter(child => child.discriminator === 'host') as HostBranch[]
-)
+const treeBranches = computed(() => branch.children.filter(child => child.discriminator === 'host') as HostBranch[])
 
-const vmLeaves = computed(() => props.branch.children.filter(child => child.discriminator === 'vm') as VmLeaf[])
+const vmLeaves = computed(() => branch.children.filter(child => child.discriminator === 'vm') as VmLeaf[])
 
 const runningVmsCount = computed(() =>
-  runningVms.value.reduce((vmCount, runningVm) => (runningVm.$pool === props.branch.data.id ? vmCount + 1 : vmCount), 0)
+  runningVms.value.reduce((vmCount, runningVm) => (runningVm.$pool === branch.data.id ? vmCount + 1 : vmCount), 0)
 )
 </script>